resource engines: move engines into separate package
Now that we have an interface for resource engines, we want to move them
into their own package so that the code's a bit more organized.
